Reuse asio's hack in other places
